[ar71xx-*] Remove watchdog driver from all builds
This change is temporarily done in Swift release branch to
enable stability test to proceed without watchdog running.

When CPU is 100% loaded using traffic, watchdog kicks in and
kills the CPU hogger, which is causing silent reboot. This
temporary change will disable watchdog to allow stability test to
continue to uncover other issues and once original issue of hogging
CPU gets fixed, this change will be reverted.

THIS CHANGE SHOULD NOT BE PROPAGATED TO ANY OTHER BRANCH.
It is a temporary change.

Change-Id: I2dea55f59a9f9e242e988966d50cc6d4fe4fb63e
Signed-off-by: Naresh Kumar Mehta <naresh@codeaurora.org>